% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APinR.R
\name{AP}
\alias{AP}
\title{Run AP commands in R wrapper function}
\usage{
AP(command)
}
\arguments{
\item{command}{command you want to pass to AP}
}
\value{
TODO
}
\description{
Tries to pass a command to AP or else return AP not installed error.
}
\examples{
call_AP("help")
}
