% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APinR.R
\name{ap_get_audio_files}
\alias{ap_get_audio_files}
\title{This function is called in \code{ap_prepare}. It gets the files, builds, and runs
the AP commands.}
\usage{
ap_get_audio_files(dirs, base_output_directory)
}
\arguments{
\item{dirs}{Specified in AP_prepare: where audio files are stored.}

\item{base_output_directory}{Specified in AP_prepare: where output files
should be stored.}
}
\description{
This function is called in \code{ap_prepare}. It gets the files, builds, and runs
the AP commands.
}
\keyword{internal}
