% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APinR.R
\name{ap_main}
\alias{ap_main}
\title{Run AP commands in R wrapper function}
\usage{
ap_main(command)
}
\arguments{
\item{command}{Character string; Command to pass to AP}
}
\value{
Returns the output of the AP command in the console.
}
\description{
Tries to pass a command to AP or else return AP not installed error.
\itemize{
\item modify to #.Platform$OS.type unix vs windows
\item A2O files have information in the names so 1 directory per file is self
explanatory for now.
}
}
\examples{
ap_main("help")

}
