% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APinR.R
\name{AP_prepare}
\alias{AP_prepare}
\title{This function runs AP analysis}
\usage{
AP_prepare(path_to_audio, path_to_output)
}
\arguments{
\item{path_to_audio}{the path to your audio files}

\item{path_to_output}{the output directory name that you would like to create and store the results in}
}
\value{
returns AP generated files
}
\description{
Generates acoustic indices using AP with Towsey config, for all `flac` files detected in the supplied path.
}
\examples{
AP_prepare("data/my_audio", "output/indices")
}
