% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_extraction.R
\name{find_matching_audio}
\alias{find_matching_audio}
\title{Uses the names in list_out$folder_path to find the corresponding audio files
for images that have been used to create crops. Relies on audio files having
the same name as the folders created by AP (which is default behaviour)}
\usage{
find_matching_audio(list_out, audio_source_dir)
}
\description{
Uses the names in list_out$folder_path to find the corresponding audio files
for images that have been used to create crops. Relies on audio files having
the same name as the folders created by AP (which is default behaviour)
}
