% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APinR.R
\name{find_program}
\alias{find_program}
\title{Runs \code{which} package}
\usage{
find_program(package)
}
\arguments{
\item{package}{name of package to check}
}
\description{
Check if a package is available on path or not.
}
\examples{
find_program("ffmpeg")

}
