% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APinR.R
\name{get_audio_files}
\alias{get_audio_files}
\title{Get audio files and run AP analysis}
\usage{
get_audio_files(dirs, base_output_directory)
}
\arguments{
\item{dirs}{specified in AP_prepare: where audio files are stored}

\item{base_output_directory}{specified in AP_prepare: where output files should be stored}
}
\description{
This function is called in AP_prepare. It gets the files, builds, and runs the AP commands
}
