% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_data_path}
\alias{get_data_path}
\title{Get data path}
\usage{
get_data_path(...)
}
\arguments{
\item{...}{enter your directory or multiple directories}
}
\value{
A full path
}
\description{
Enter a directory relative to the current working directly, and return the
full path.
}
\examples{
get_data_path()
get_data_path("data/sounds")

}
\keyword{internal}
