% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_plots.R
\name{hime_indices}
\alias{hime_indices}
\title{hime_indices function}
\usage{
hime_indices(
  total_indices,
  complete_ts,
  himedatapath,
  X.sites,
  X.month_id,
  figures
)
}
\arguments{
\item{total_indices}{the three indices}

\item{complete_ts}{the timeseries data frame}

\item{himedatapath}{the path to hime data}

\item{X.sites}{site ID}

\item{X.month_id}{month ID}

\item{figures}{where the figures will be stored}
}
\description{
matches hime data to time series indices data. generates plots (one per index per site per month)
}
