% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hime_processing.R
\name{hime_processing}
\alias{hime_processing}
\title{Hime processing}
\usage{
hime_processing(himeoutput)
}
\arguments{
\item{himeoutput}{the output folder}
}
\value{
creates a (whatever you called your himeoutput)-clean folder with cleaned hime txt files
}
\description{
Hime processing
}
