% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_plots.R
\name{indicestimeplots}
\alias{indicestimeplots}
\title{indicestimeplots function}
\usage{
indicestimeplots(complete_ts, pathid, siteID, total_indices, Y.month_id)
}
\arguments{
\item{complete_ts}{the timeseries data frame}

\item{pathid}{where output figures are stored}

\item{siteID}{the study site ID}
}
\description{
plots time series of all 3 indices
}
