% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_repeated.R
\name{iteration1}
\alias{iteration1}
\title{Iteration function}
\usage{
iteration1(motif_results_df)
}
\arguments{
\item{motif_results_df}{a motif results df}
}
\value{
returns a motif results df
}
\description{
iteration1 will be used when cleaning up the motifs to remove repetition. This function is assigning the word “repeated” to motifs that overlap.
}
