% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifR-package.R
\docType{package}
\name{MIMiCS-package}
\alias{MIMiCS}
\alias{MIMiCS-package}
\title{MIMiCS: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
