% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mimics_process.R
\name{mimics_check_indices_data}
\alias{mimics_check_indices_data}
\title{Check if indices data is of valid input format.}
\usage{
mimics_check_indices_data(my_indices_data)
}
\description{
Helper function for \code{mimics_process()}.
}
\keyword{internal}
