% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mimics_process.R
\name{mimics_process}
\alias{mimics_process}
\title{Combine motif and time series data and generate outputs}
\usage{
mimics_process(my_indices_data, outputfigures, himepath, threshold)
}
\arguments{
\item{my_indices_data}{Data frame produced by \code{time_series()} as input}

\item{outputfigures}{Character string; Path to folder where output figures
will be saved}

\item{himepath}{Character string; Path to folder where hime motif results
were saved when using the \code{run_hime()} function.}

\item{threshold}{Numeric between 0 and 1; Threshold for removing overlapping
motifs. Default is 0.9.}
}
\description{
Combine motif and time series data and generate outputs
}
