% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_plots.R
\name{motif_plots}
\alias{motif_plots}
\title{Motif plots}
\usage{
motif_plots(data_indices_all, outputfigures, himedatapath)
}
\arguments{
\item{data_indices_all}{the data frame returned by motifR::time_series}

\item{outputfigures}{a relative folder path where you want to store figure output}

\item{himedatapath}{the relative path to your hime data cleaned folder}
}
\description{
This function will use your indices data, and hime data, to generate plots
and combined motif/index data frames (.csv files)
}
