% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series.R
\name{read_indices}
\alias{read_indices}
\title{Read indices function}
\usage{
read_indices(files)
}
\arguments{
\item{files}{your files}
}
\description{
Combines ACI, ENT, TE indices data into a single data frame.
Uses / generates metadata based on the file/ folder names
}
\examples{
read_indices(files)
}
