% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_repeated.R
\name{remove_repeated}
\alias{remove_repeated}
\title{remove repeated}
\usage{
remove_repeated(motif_results_df)
}
\arguments{
\item{motif_results_df}{a motif results df}
}
\description{
Now this function actually cleans up the motifs.
remove_repeated runs the iteration 3 times so it removes
all the possible repetitions and overlaps and
removes motifs assigned as repeated.
}
