% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hime.R
\name{run_hime}
\alias{run_hime}
\title{Run hime}
\usage{
run_hime(
  timeseriesdata,
  himeoutput,
  himepath = file.path(getwd(), "HIME-master/bin/HIME_release.jar")
)
}
\arguments{
\item{timeseriesdata}{the folder you output the timeseries data to}

\item{himeoutput}{the folder you want to store the hime output}

\item{himepath}{is the path to your hime_release.jar. default is a folder in your WD}
}
\description{
Run hime
}
\examples{
run_hime(run_hime(timeseriesdata, himeoutput, himepath = "~/HIME/bin/HIME_release.jar"))
}
