% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series.R
\name{subset_indices}
\alias{subset_indices}
\title{Subset indices timeseries data for input to HIME}
\usage{
subset_indices(indicesdata, subsetlist, timeseriesfolder)
}
\arguments{
\item{indicesdata}{indices data}

\item{subsetlist}{list to subset by indices}

\item{timeseriesfolder}{where the output will be stored}
}
\description{
The output here should be: files subset by geo/site ID (each sensor location), by month, by index
}
\examples{
subset_indices(indicesdata, subsetlist, timeseriesfolder)
}
