% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series.R
\name{time_series}
\alias{time_series}
\title{Generate acoustic indices time series}
\usage{
time_series(path_to_indices, path_to_output)
}
\arguments{
\item{path_to_indices}{this is where your indices are saved}

\item{path_to_output}{this is where you would like the time series data saved}
}
\value{
Data frame
}
\description{
Output is a folder which contains .csv and .txt files of time series data for each acoustic index.
One file per index, per site, per month
}
\examples{
time_series(
  indicesfolder = "output/indices-output-ap",
  outputfolder = "output/timeseries"
)
}
