% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series.R
\name{timeseries_prep}
\alias{timeseries_prep}
\title{Time series prep}
\usage{
timeseries_prep(indicesdata, outputfolder)
}
\arguments{
\item{outputfolder}{}
}
\description{
Creates the output folder and creates the list that will be used for subsetting
}
