% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APinR.R
\name{ap_prepare}
\alias{ap_prepare}
\title{Generate acoustic indices using Analysis Programs}
\usage{
ap_prepare(path_to_audio, path_to_output)
}
\arguments{
\item{path_to_audio}{Character string; The path to your audio files.}

\item{path_to_output}{Character string; The output directory name that you
would like to use to store the results in. Created if it doesn't exist.}
}
\value{
Returns AP generated files in the output directory.
}
\description{
Generates acoustic indices using AP with Towsey config, for all \code{flac} files
detected in the supplied path.
}
\examples{
ap_prepare("data/my_audio", "output/indices")

}
