% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_program.R
\name{find_program}
\alias{find_program}
\title{Runs `which` package}
\usage{
find_program(package)
}
\arguments{
\item{package}{name of package to check}
}
\value{
todo
}
\description{
To check if a package is available on path or not
}
\examples{
find_program("ffmpeg")
}
