% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series.R
\name{time_series}
\alias{time_series}
\title{Generate acoustic indices time series}
\usage{
time_series(indicesfolder, outputfolder)
}
\arguments{
\item{indicesfolder}{this is where your indices are saved}

\item{outputfolder}{this is where you would like the timeseries data saved}
}
\value{
a dataframe
}
\description{
Output is a folder which contains .csv and .txt files of time series data for each acoustic index.
One file per index, per site, per month
}
\examples{
time_series(indicesfolder = "output/indices-output-ap", outputfolder = "output/timeseries")
}
